backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]  # Updated to poetry-core
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "A FastAPI application"
packages = [
    { include = "app", from = "." }  # Critical addition
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
langchain-core = "^0.3.0"
temporalio = "^1.5.0"
supabase = "^2.8.0"
docker = "^7.0.0"
redis = "^5.0.0"
jinja2 = "^3.1.6"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
isort = "^5.13.0"
